add annos, cleanup sc2
