+++ components/infobox/wikis/starcraft/infobox_skill_custom.lua
@@ -27,6 +27,8 @@ local CustomInjector = Class.new(Injector)
 
 local _args
 
+---@param frame Frame
+---@return Html
 function CustomSkill.run(frame)
 	local customSkill = Skill(frame)
 	customSkill.createWidgetInjector = CustomSkill.createWidgetInjector
@@ -36,6 +38,8 @@ function CustomSkill.run(frame)
 	return customSkill:createInfobox()
 end
 
+---@param widgets Widget[]
+---@return Widget[]
 function CustomInjector:addCustomCells(widgets)
 	return {
 		Cell{name = 'Area of Effect', content = {_args.area}},
@@ -46,6 +50,9 @@ function CustomInjector:addCustomCells(widgets)
 	}
 end
 
+---@param id string
+---@param widgets Widget[]
+---@return Widget[]
 function CustomInjector:parse(id, widgets)
 	if id == 'cost' then
 		return {Cell{name = 'Cost', content = {CustomSkill:getCostDisplay()}}}
@@ -60,10 +67,12 @@ function CustomInjector:parse(id, widgets)
 	return widgets
 end
 
+---@return WidgetInjector
 function CustomSkill:createWidgetInjector()
 	return CustomInjector()
 end
 
+---@return string?
 function CustomSkill:getResearchCost()
 	if String.isEmpty(_args.from) then
 		return nil
@@ -78,6 +87,7 @@ function CustomSkill:getResearchCost()
 	}
 end
 
+---@return string
 function CustomSkill:getResearchFrom()
 	if String.isEmpty(_args.from) then
 		return 'No research needed'
@@ -88,6 +98,7 @@ function CustomSkill:getResearchFrom()
 	end
 end
 
+---@return string?
 function CustomSkill:getResearchHotkey()
 	if String.isEmpty(_args.from) then
 		return
@@ -96,6 +107,8 @@ function CustomSkill:getResearchHotkey()
 	return Hotkeys.hotkey(_args.rshortcut)
 end
 
+---@param args table
+---@return string[]
 function CustomSkill:getCategories(args)
 	local skill = args.informationType .. 's'
 	local categories = {skill}
@@ -107,6 +120,7 @@ function CustomSkill:getCategories(args)
 	return categories
 end
 
+---@return string?
 function CustomSkill:getHotkeys()
 	if not String.isEmpty(_args.shortcut) then
 		if not String.isEmpty(_args.shortcut2) then
@@ -117,6 +131,7 @@ function CustomSkill:getHotkeys()
 	end
 end
 
+---@return string
 function CustomSkill:getCostDisplay()
 	local energy = tonumber(_args.energy) or 0
 	return ENERGY_ICON .. '&nbsp;' .. energy
